Release 1.11.2 (#675)
