+++ src/components/Messages/MessageText/markdown.js
@@ -1,4 +1,3 @@
-import { sanitize } from 'dompurify';
 import MarkdownIt from 'markdown-it';
 
 
@@ -81,4 +80,4 @@ md.use((md) => {
 	});
 });
 
-export const renderMarkdown = (...args) => sanitize(md.render(...args), { ADD_ATTR: ['target', 'rel'] });
+export const renderMarkdown = (...args) => md.render(...args);
