+++ src/components/helpers.js
@@ -1,4 +1,5 @@
 import parseISO from 'date-fns/parseISO';
+import mem from 'mem';
 import { Component } from 'preact';
 
 import { Livechat, useSsl } from '../api';
@@ -262,3 +263,23 @@ export const resolveDate = (dateInput) => {
 		}
 	}
 };
+
+const escapeMap = {
+	'&': '&amp;',
+	'<': '&lt;',
+	'>': '&gt;',
+	'"': '&quot;',
+	'\'': '&#x27;',
+	'`': '&#x60;',
+};
+
+const escapeRegex = new RegExp(`(?:${ Object.keys(escapeMap).join('|') })`, 'g');
+
+const escapeHtml = mem(
+	(string) => string.replace(escapeRegex, (match) => escapeMap[match]),
+);
+
+export const parse = (plainText) =>
+	[{ plain: plainText }]
+		.map(({ plain, html }) => (plain ? escapeHtml(plain) : html || ''))
+		.join('');
